\docType{data}
\name{data.Y.Hopx}
\alias{data.Y.Hopx}
\title{Data set compiling gene expression levels (HOPX gene)  for 29 rats.}
\format{A \code{29 x 4} data frame containing the gene expression levels for 29
  rats (in rows) in 4 different tissues (in columns): ADR, FAT, Heart, Kidney}
\description{
  This data set contains gene expression levels for gene
  Hopx from 4 tissues (Adrenal gland, Fat, Heart,
  Kidney) in 29 rats.
}
\keyword{datasets}

