\name{example.as.ESS.object}

\alias{example.as.ESS.object}

\title{Function creating an \code{ESS} object from the example files
  contained in the package}


\usage{example.as.ESS.object()}

\description{ This functions returns an \code{ESS} object constructed
  from the example files provided within the package.}

\value{ This functions returns an \code{ESS} object constructed from
  the example files provided within the package. The object contains
  fields as listed in \code{\link{as.ESS.object}}. This function is
  used throughout the documentation to generate the example \code{ESS}
  object used to illustrate the functions embedded the \code{\link{R2GUESS}}
  package.}

\seealso{ \code{\link{as.ESS.object}}}