\name{get.best.models}
\alias{get.best.models}
\title{Internal function used to read in R the result of the best model}
\usage{
  get.best.models(path.out, path.in = NULL, file.in, label.X = NULL, p, MAP.file = NULL)
}
\description{
  Internal function used to read in R the result of the
  best model
}
\keyword{internal}

