\docType{data} \name{MAP.file} \alias{MAP.file} \title{MAP file
  describing genotypes from the rats experiment.}  \format{A data
  frame with 770 rows and 3 columns} \description{ This data set
  contains the annotation information for the 770 SNPs used in the
  example. SNPs are described in rows (in the same order as they appear
  in \code{\link{data.X}}), and the following information is provided:
  the SNP rs name (\code{SNPName}), the Chromosome it belongs to
  (\code{Chr}), and the physical position on the chromosome (\code{Posn})}


\keyword{datasets}

\seealso{ \code{\link{data.X}}}