\name{wrapper}
\alias{wrapper}
\title{Internal function used for the parallel computation on the permutation procedure for the calibration of the FDR threshold
regarding the significance of the MPPI}
\usage{
  wrapper(x, root.file.output, object, path.input,
    path.output, Y = Y,start.counter)
}
\description{
  Internal function used for the parallel computation on
  the permutation procedure for the calibration of the FDR
  threshold regarding the significance of the MPPI
}
\keyword{internal}

